+++ src/app/meetings/[id]/_components/MeetingHeaderSection.tsx
@@ -15,7 +15,7 @@ import {
   DropdownMenuTrigger,
 } from "@/components/ui/Dropdown";
 import { DeleteModal } from "@/components/modal/DeleteModal";
-import { HeartIcon } from "@/components/icon/HeartIcon";
+import { HeartIcon } from "@/components/ui/HeartIcon";
 import FallbackImage from "@/components/img/FallbackImage";
 import { useLoginModalStore } from "@/store/useLoginModalStore";
 import {
