+++ src/auth/fetcher.client.ts
@@ -24,6 +24,7 @@ axiosInstance.interceptors.response.use(
     ) {
       // 이미 로그인 페이지에 있으면 리다이렉트하지 않음 (무한 루프 방지)
       if (window.location.pathname !== "/login") {
+        await fetch("/api/auth/logout", { method: "POST" }).catch(() => {});
         window.location.href = "/login";
         // 에러 객체에 속성 추가
         const errorWithFlag = error;
