+++ src/hooks/queries/useMeetingList.ts
@@ -7,10 +7,19 @@ import type {
   GetMeetingListParams,
   SortOrder,
   MeetingSortBy,
+  JoinedMeeting
 } from "@/types";
 import { getNextPageParam } from "@/lib/pagination";
+import {
+  InfiniteData,
+  QueryKey,
+  useMutation,
+  useQueryClient,
+} from "@tanstack/react-query";
+import { deleteFavorites, updateFavorites } from "@/api/client";
 import { QUERY_KEYS } from "@/constants/queryKey";
-import type { QueryKey } from "@tanstack/react-query";
+import { useOptimisticMutation } from "@/hooks/useOptimisticUpdate";
+
 
 interface GetMeetingsProps {
   type?: string;
@@ -95,3 +104,48 @@ export const useJoinedMeetingList = (enabled = true): InfiniteListResult => {
     favoriteQueryKey: QUERY_KEYS.meetings.joined,
   };
 };
+
+
+
+export const useMeetingFavoriteMutation = (
+  queryKey: QueryKey = QUERY_KEYS.meetings.joined,
+) => {
+  const queryClient = useQueryClient();
+
+  const mutation = useMutation({
+    mutationFn: async (meeting: Pick<JoinedMeeting, "id" | "isFavorited">) => {
+      if (meeting.isFavorited) {
+        await deleteFavorites(meeting.id);
+        return;
+      }
+      await updateFavorites(meeting.id);
+    },
+    ...useOptimisticMutation<
+      InfiniteData<JoinedMeetingsResponse>,
+      Pick<JoinedMeeting, "id" | "isFavorited">
+    >(queryClient, {
+      queryKey,
+      updater: (oldData, meeting) => ({
+        ...oldData,
+        pages: oldData.pages.map((page) => ({
+          ...page,
+          data: page.data.map((item) =>
+            item.id === meeting.id
+              ? { ...item, isFavorited: !item.isFavorited }
+              : item,
+          ),
+        })),
+      }),
+      invalidateKeys: [QUERY_KEYS.meetings.root, QUERY_KEYS.favorites.root],
+      onErrorMessage: "즐겨찾기 처리에 실패했습니다.",
+    }),
+  });
+
+  return {
+    toggleFavorite: mutation.mutate,
+    toggleFavoriteAsync: mutation.mutateAsync,
+    isPending: mutation.isPending,
+    error: mutation.error,
+  };
+};
+
